finish test jest
